Day 2
I create an app that allows a user to input information about themselves by using input variables and the print function. I also stumbled into figuring out a way to embed the users input variable within a question as if the user is having a conversation with the store attendant.